save button in extra playground controls when editing a model
